barcode example working

+++ packages/barcode/android/src/main/java/com/reactnativecamera/barcode/RectOfInterest.kt
@@ -1,3 +1,3 @@
 package com.reactnativecamera.barcode
 
-class RectOfInterest(val x: Float, val y: Float, val width: Float, val height: Float)
+class RectOfInterest(val x: Int, val y: Int, val width: Int, val height: Int)
